refactor(all): moved common properties in config class
