+++ src/base/config.ts
@@ -1,5 +1,5 @@
 import { Component, Project } from 'projen';
-// import { ProjectTypes } from '../types/types';
+import { ConfigFile, Settings } from '../types/types';
 
 /**
  * Base class for creating and managing project configurations.
@@ -13,6 +13,87 @@ export class Config extends Component {
    */
   constructor(project: Project) {
     super(project);
+  }
+
+  /**
+   * Gets the additional development dependencies required for configuration.
+   *
+   * @returns A list of package names with version specifications.
+   */
+  protected get additionalDevDependencies(): string[] {
+    return [];
+  }
+
+  /**
+   * Gets the additional peer dependencies required for configuration.
+   *
+   * @returns A list of package names with version specifications.
+   */
+  protected get additionalPeerDependencies(): string[] {
+    return [];
+  }
+
+  /**
+   * Gets the additional dependencies required for configuration.
+   *
+   * @returns A list of package names with version specifications.
+   */
+  protected get additionalDependencies(): string[] {
+    return [];
+  }
+
+  /**
+   * Gets the additional settings to be added to the project's configuration.
+   *
+   * @returns A settings object to be merged into the project's settings.
+   */
+  protected get additionalSettings(): Settings {
+    return {};
+  }
+
+  /**
+   * Gets the additional npm scripts to be added to the project's configuration.
+   *
+   * @returns A record of script names and their corresponding commands.
+   */
+  protected get additionalScripts(): Record<string, string> {
+    return {};
+  }
+
+  /**
+   * Gets additional ignore patterns to be added to the project's ignore configuration.
+   *
+   * @returns A list of ignore patterns.
+   */
+  protected get additionalIgnorePatterns(): string[] {
+    return [];
+  }
+
+  /**
+   * Gets additional attributes patterns to be added to the project's attributes configuration.
+   *
+   * @returns A list of attibute patterns.
+   */
+  protected get additionalAttributesPatterns(): string[] {
+    return [];
+  }
+
+  /**
+   * Gets the configuration file content.
+   *
+   * @returns An object where the key is the filename and the value is an array of file lines.
+   */
+  protected get configFile(): ConfigFile {
+    return {};
+  }
+
+  /**
+   * Gets the ignore file content.
+   *
+   * @returns An object where the key is the filename and the value is an array of file lines.
+   */
+  protected get ignoreFile(): Record<string, string[]> {
+    return {};
   }
 
   /**
