+++ src/base/devcontainer.ts
@@ -1,7 +1,8 @@
 import { JsonFile } from 'projen';
 import { Config } from './config';
-import { ProjectTypes, Settings } from '../types/types';
+import { ProjectTypes } from '../types/project';
 import { isValidProject } from '../utils';
+import { Settings } from '../types/types';
 
 /**
  * Base class for implementing all relevant DevContainer configuration.
@@ -9,23 +10,13 @@ import { isValidProject } from '../utils';
  * This class acts as a base for handling DevContainer configuration within projects.
  */
 export class DevContainerConfigBase extends Config {
-  /**
-   * Gets the additional npm scripts to be added to the project's configuration.
-   *
-   * @returns A record of script names and their corresponding commands.
-   */
-  protected get additionalScripts(): Record<string, string> {
+  protected override get additionalScripts(): Record<string, string> {
     return {
       'install-dependencies': 'npm install',
     };
   }
 
-  /**
-   * Gets the config file to be added to the project's configuration.
-   *
-   * @returns A record of the having the path to the file as key and the content as value.
-   */
-  protected get configFile(): Settings {
+  protected override get configFile(): Settings {
     return {
       '.devcontainer.json': {
         image: 'mcr.microsoft.com/devcontainers/typescript-node:1-20-bullseye',
@@ -79,12 +70,7 @@ export class DevContainerConfigBase extends Config {
     };
   }
 
-  /**
-   * Gets additional ignore patterns to be added to the project's ignore configuration.
-   *
-   * @returns A list of ignore patterns.
-   */
-  protected get additionalIgnorePatterns(): string[] {
+  protected override get additionalIgnorePatterns(): string[] {
     const filePath: string = Object.keys(this.configFile)[0];
     return [`/${filePath}`];
   }
