+++ src/cap-service/eslint.ts
@@ -4,7 +4,7 @@ import { EsLintConfigBase } from '../base';
  * Implementing all relevant EsLint configuration for the CapService project.
  */
 export class EsLintConfigCapService extends EsLintConfigBase {
-  protected get standardIgnorePatterns(): string[] {
+  protected override get standardIgnorePatterns(): string[] {
     return ['test/', 'dist/', 'gen/', '@cds-models', 'default-env.js', '@dispatcher'];
   }
 }
