+++ src/cap-service/jest.ts
@@ -1,38 +1,24 @@
 import { JestConfigBase } from '../base/jest';
-import { ProjectTypes, Settings } from '../types/types';
+import { ProjectTypes } from '../types/project';
+import { Settings } from '../types/types';
 import { isValidProject } from '../utils';
 
 /**
  * Implementing all relevant Jest configuration for the CapService project.
  */
 export class JestConfigCapService extends JestConfigBase {
-  /**
-   * Gets the additional development dependencies required for configuration.
-   *
-   * @returns A list of package names with version specifications.
-   */
-  protected get additionalDevDependencies(): string[] {
+  protected override get additionalDevDependencies(): string[] {
     return ['@types/jest@^29.5.14', 'jest@^29.7.0', 'jest-junit@^16.0.0', 'ts-jest@^29.2.5'];
   }
 
-  /**
-   * Gets the additional npm scripts to be added to the project's configuration.
-   *
-   * @returns A record of script names and their corresponding commands.
-   */
-  protected get additionalScripts(): Record<string, string> {
+  protected override get additionalScripts(): Record<string, string> {
     return {
       test: 'jest --passWithNoTests --updateSnapshot',
       'test:watch': 'jest --watch',
     };
   }
 
-  /**
-   * Gets the additional settings to be added to the project's configuration.
-   *
-   * @returns A settings object to be merged into the project's settings.
-   */
-  protected get additionalSettings(): Settings {
+  protected override get additionalSettings(): Settings {
     return {
       jest: {
         coverageProvider: 'v8',
@@ -70,12 +56,7 @@ export class JestConfigCapService extends JestConfigBase {
     };
   }
 
-  /**
-   * Gets additional ignore patterns to be added to the project's ignore configuration.
-   *
-   * @returns A list of ignore patterns.
-   */
-  protected get additionalIgnorePatterns(): string[] {
+  protected override get additionalIgnorePatterns(): string[] {
     return ['/coverage/', '/test-reports/', '/junit.xml'];
   }
