+++ src/github-action/commitlint.ts
@@ -5,7 +5,7 @@ import { Settings } from '../types';
  * Implementing all relevant CommitLint configuration for the GitHubAction project.
  */
 export class CommitLintConfigGitHubAction extends CommitLintConfigBase {
-  protected get additionalSettings(): Settings {
+  protected override get additionalSettings(): Settings {
     return {
       'lint-staged': {
         '**/*.{yml,yaml}': ['npm run prettier'],
