+++ src/github-action/npm.ts
@@ -6,23 +6,13 @@ import { Settings } from '../types';
  * Implementing all relevant NPM configuration for the GitHubAction project.
  */
 export class NpmConfigGitHubAction extends NpmConfigBase {
-  /**
-   * Gets the additional settings to be added to the project's configuration.
-   *
-   * @returns A settings object to be merged into the project's settings.
-   */
-  private get additionalSettings(): Settings {
+  protected override get additionalSettings(): Settings {
     return {
       files: ['action.yml', 'README.md'],
     };
   }
 
-  /**
-   * Gets additional ignore patterns to be added to the project's ignore configuration.
-   *
-   * @returns A list of ignore patterns.
-   */
-  private get additionalIgnorePatterns(): string[] {
+  protected override get additionalIgnorePatterns(): string[] {
     return ['docs/'];
   }
