+++ src/types/index.ts
@@ -1,1 +1,2 @@
+export * from './project';
 export * from './types';
